+++ Requester/.env
@@ -1,0 +1,4 @@
+TOKEN=<token>
+RAVENVERSE_URL=http://0.0.0.0
+RAVENVERSE_FTP_HOST=0.0.0.0
+RAVENVERSE_FTP_URL=0.0.0.0
