+++ Requester/create_distributed_cnn.py
@@ -4,7 +4,7 @@ load_dotenv()
 import os
 import ravop as R
 
-from examples import cnn
+from .examples import cnn
 
 # Initialize and create graph
 R.initialize(ravenverse_token=os.environ.get("TOKEN"))
