+++ Requester/create_distributed_knn.py
@@ -3,7 +3,7 @@ load_dotenv()
 
 import os
 import ravop as R
-from examples import knn
+from .examples import knn
 
 # Initialize and create graph
 R.initialize(ravenverse_token=os.environ.get("TOKEN"))
